feat: fetch()를 사용하여 데이터 불러와 처리하기
