--- data/data.js
@@ -1,91 +1,0 @@
-export const data = {
-  colums: [
-    {
-      id: "toDo",
-      count: 0,
-      cardList: [
-        {
-          title: "GitHub 공부하기",
-          content: "add, commit, push",
-          author: "author by web",
-        },
-        {
-          title: "블로그에 포스팅할 것",
-          content: "* GitHub 공부내용 \n * 모던 자바스크립트 1장 공부내용",
-          author: "author by web",
-        },
-        {
-          title: "GitHub 공부하기",
-          content: "add, commit, push",
-          author: "author by web",
-        },
-      ],
-    },
-    {
-      id: "doing",
-      count: 0,
-      cardList: [
-        {
-          title: "HTML/CSS 공부하기",
-          content: "input 태그 실습",
-          author: "author by web",
-        },
-      ],
-    },
-    {
-      id: "done",
-      count: 0,
-      cardList: [
-        {
-          title: "HTML/CSS 공부하기",
-          content: "input 태그 실습",
-          author: "author by web",
-        },
-      ],
-    },
-  ],
-  historyList: [
-    {
-      photo: "./images/sam.png",
-      userName: "@멋진삼",
-      textBody:
-        "블로그에 포스팅할 것 <span> 을(를)</span> 하고있는 일<span> 에서 </span><br />해야할 일 <span> 으로 </span> 이동 <span> 하였습니다. </span>",
-      timeStamp: "3분 전",
-    },
-    {
-      photo: "./images/sam.png",
-      userName: "@멋진삼",
-      textBody:
-        "GitHub 공부하기 <span> 을(를)</span> 변경<span> 하였습니다. </span>",
-      timeStamp: "10분 전",
-    },
-    {
-      photo: "./images/sam.png",
-      userName: "@멋진삼",
-      textBody:
-        "HTML/CSS 공부하기 <span> 을(를)</span> 해야할 일 <span> 에서 </span><br />하고 있는 일 <span> 으로 </span> 이동 <span> 하였습니다. </span>",
-      timeStamp: "18분 전",
-    },
-    {
-      photo: "./images/sam.png",
-      userName: "@멋진삼",
-      textBody:
-        "HTML/CSS 공부하기 <span> 을(를)</span> 해야할 일<span> 에서 </span><br />등록 <span> 하였습니다. </span>",
-      timeStamp: "1시간 전",
-    },
-    {
-      photo: "./images/sam.png",
-      userName: "@멋진삼",
-      textBody:
-        "HTML/CSS 공부하기 <span> 을(를)</span> 해야할 일<span> 에서 </span><br />등록 <span> 하였습니다. </span>",
-      timeStamp: "1시간 전",
-    },
-    {
-      photo: "./images/sam.png",
-      userName: "@멋진삼",
-      textBody:
-        "HTML/CSS 공부하기 <span> 을(를)</span> 해야할 일<span> 에서 </span><br />등록 <span> 하였습니다. </span>",
-      timeStamp: "1시간 전",
-    },
-  ],
-};
