+++ index.html
@@ -185,7 +185,6 @@
     </div>
     <script src="./eventHandler/historyList.js"></script>
     <script src="./eventHandler/textarea.js"></script>
-    <script type="module" src="./eventHandler/cardView.js"></script>
-    <script type="module" src="./eventHandler/historyView.js"></script>
+    <script type="module" src="./eventHandler/view.js"></script>
   </body>
 </html>
